Feat : 공공데이터 호출 시 서킷브레이커 추가
event-service, sbike-service 서킷브레이커 추가

+++ sbike-service/src/main/java/com/sth/sbikeservice/model/dto/SbikeDTO.java
@@ -46,7 +46,7 @@ public class SbikeDTO {
 
     public Sbike toEntity() {
         return Sbike.builder()
-                .id(id)
+                .id((int) id)
                 .stationId(stationId)
                 .rackTotCnt(rackTotCnt)
                 .stationName(stationName)
